# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Expense Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /platforms:
    get:
      operationId: listPlatforms
      tags: [platforms]
      summary: List all AI platforms
      responses:
        "200":
          description: List of platforms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Platform"
    post:
      operationId: createPlatform
      tags: [platforms]
      summary: Create a new platform
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatformInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Platform"
  /platforms/{id}:
    get:
      operationId: getPlatform
      tags: [platforms]
      summary: Get a platform by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Platform
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Platform"
    patch:
      operationId: updatePlatform
      tags: [platforms]
      summary: Update a platform
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatformUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Platform"
    delete:
      operationId: deletePlatform
      tags: [platforms]
      summary: Delete a platform
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    get:
      operationId: getExpense
      tags: [expenses]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    patch:
      operationId: updateExpense
      tags: [expenses]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /credits:
    get:
      operationId: listCreditPurchases
      tags: [credits]
      summary: List all credit top-up purchases
      responses:
        "200":
          description: List of credit purchases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CreditPurchase"
    post:
      operationId: createCreditPurchase
      tags: [credits]
      summary: Log a credit top-up purchase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreditPurchaseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditPurchase"
  /credits/{id}:
    patch:
      operationId: updateCreditPurchase
      tags: [credits]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreditPurchaseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditPurchase"
    delete:
      operationId: deleteCreditPurchase
      tags: [credits]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Platform:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        website:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        apiKey:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    PlatformInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        website:
          type: string
        logoUrl:
          type: string
        category:
          type: string
        email:
          type: string
        apiKey:
          type: string
        notes:
          type: string
    PlatformUpdate:
      type: object
      properties:
        name:
          type: string
        website:
          type: string
        logoUrl:
          type: string
        category:
          type: string
        email:
          type: string
        apiKey:
          type: string
        notes:
          type: string
...
    Project:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        color:
          type: string
        status:
          type: string
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        color:
          type: string
        status:
          type: string
    Expense:
      type: object
      required: [id, amount, date, createdAt]
      properties:
        id:
          type: integer
        platformId:
          type: ["integer", "null"]
        platformName:
          type: ["string", "null"]
        projectId:
          type: ["integer", "null"]
        projectName:
          type: ["string", "null"]
        amount:
          type: number
        currency:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        date:
          type: string
        createdAt:
          type: string
    ExpenseInput:
      type: object
      required: [amount, date]
      properties:
        platformId:
          type: integer
        projectId:
          type: integer
        amount:
          type: number
        currency:
          type: string
        description:
          type: string
        category:
          type: string
        date:
          type: string
    ExpenseUpdate:
      type: object
      properties:
        platformId:
          type: integer
        projectId:
          type: integer
        amount:
          type: number
        currency:
          type: string
        description:
          type: string
        category:
          type: string
        date:
          type: string
...
    CreditPurchase:
      type: object
      required: [id, platformId, amount, purchaseDate, createdAt]
      properties:
        id:
          type: integer
        platformId:
          type: integer
        platformName:
          type: ["string", "null"]
        projectId:
          type: ["integer", "null"]
        projectName:
          type: ["string", "null"]
        amount:
          type: number
        credits:
          type: ["number", "null"]
        currency:
          type: string
        description:
          type: ["string", "null"]
        purchaseDate:
          type: string
        createdAt:
          type: string
    CreditPurchaseInput:
      type: object
      required: [platformId, amount, purchaseDate]
      properties:
        platformId:
          type: integer
        projectId:
          type: integer
        amount:
          type: number
        credits:
          type: number
        currency:
          type: string
        description:
          type: string
        purchaseDate:
          type: string
    CreditPurchaseUpdate:
      type: object
      properties:
        platformId:
          type: integer
        projectId:
          type: integer
        amount:
          type: number
        credits:
          type: number
        currency:
          type: string
        description:
          type: string
        purchaseDate:
          type: string
...